feat && refactoring: rewrite stucture of test to match new parts of testsing framework
New macroses are used and now tests are counted
and give correct error msgs, if some of them fails.

Parts of the initial testing structure has
been revisioned. Mostly commit emits goto cleanup and use loops for
initial values where it was necessary).

Also the commit gives an example of the test with more stable
conversion of the tested list to an intermediate representation.